--- src/components/days/index.js
@@ -1,4 +1,0 @@
-import {createTripDayList} from "./trip-day-list.js";
-import {createTripDay} from "./trip-day.js";
-
-export {createTripDayList, createTripDay};
