--- src/components/days/trip-day-list.js
@@ -1,1 +1,0 @@
-export const createTripDayList = () => `<ul class="trip-days"></ul>`;
