+++ src/components/event/event-day-list.js
@@ -1,0 +1,1 @@
+export const createEventDayList = () => `<ul class="trip-days"></ul>`;
