+++ src/components/event/event-day.js
@@ -1,11 +1,14 @@
 import {getDateTimeFormat, getFormatedDate} from "../../utils.js";
+import {createEventList} from "./event-list.js";
 
-export const createTripDay = (date, count) => {
+export const createEventDay = (date, count) => {
   return `<li class="trip-days__item day">
   <div class="day__info">
     <span class="day__counter">${count}</span>
     <time class="day__date" datetime="${getDateTimeFormat(date)}">${getFormatedDate(date).toLowerCase()}</time>
   </div>
+  ${createEventList()}
 </li>
+
 `;
 };
