+++ src/components/event/index.js
@@ -1,5 +1,7 @@
-import {createTripEventForm} from "./event-form.js";
+import {createEventForm} from "./event-form.js";
 import {createEventItem} from "./event-item.js";
 import {createEventList} from "./event-list.js";
+import {createEventDayList} from "./event-day-list.js";
+import {createEventDay} from "./event-day.js";
 
-export {createTripEventForm, createEventItem, createEventList};
+export {createEventForm, createEventItem, createEventList, createEventDayList, createEventDay};
