+++ src/components/index.js
@@ -1,14 +1,13 @@
-import {createTripDayList, createTripDay} from "./days/index.js";
-import {createTripEventForm, createEventItem, createEventList} from "./event/index.js";
+import {createEventForm, createEventItem, createEventList, createEventDayList, createEventDay} from "./event/index.js";
 import {createTripFilter} from "./filter/index.js";
 import {createTripInfo} from "./info/index.js";
 import {createTripMenu} from "./menu/index.js";
 import {createTripSortForm} from "./sort-form/index.js";
 
 export {
-  createTripDayList,
-  createTripDay,
-  createTripEventForm,
+  createEventDayList,
+  createEventDay,
+  createEventForm,
   createEventItem,
   createEventList,
   createTripFilter,
