+++ src/components/info/trip-info-routes.js
@@ -1,0 +1,14 @@
+import {getFormatedDate} from '../../utils.js';
+
+export const createRoutesInfo = (routes) => {
+  const startRoute = routes[1].destination;
+  const middleRoute = routes[Math.floor(routes.length) / 2].destination;
+  const endRoute = routes[routes.length - 1].destination;
+  const startDate = getFormatedDate(routes[1].tripDates.start);
+  const endDate = getFormatedDate(routes[routes.length - 1].tripDates.end);
+  return `
+  <div class="trip-info__main">
+    <h1 class="trip-info__title">${startRoute} — ${middleRoute} — ${endRoute}</h1>
+    <p class="trip-info__dates">${startDate}&nbsp;—&nbsp;${endDate}</p>
+  </div>`;
+};
